+++ plugin/withExpoNumerator.js
@@ -1,5 +1,6 @@
-function withExpoNumerator(config) {
+function withExpoNumerator(config, _options) {
   return config;
 }
 
 module.exports = withExpoNumerator;
+module.exports.default = withExpoNumerator;